chore: renaming work component file
