+++ src/pages/join-us.js
@@ -4,7 +4,7 @@ import { StaticQuery, graphql } from 'gatsby'
 import Layout from '../components/layout'
 import ViewPositions from '../components/JoinUs/ViewOpenPositions'
 import Learning from '../components/JoinUs/Learning'
-import Work from '../components/JoinUs/Workplace'
+import Work from '../components/JoinUs/Work'
 import OSS from '../components/JoinUs/OpenSource'
 import Perks from '../components/JoinUs/Perks'
 import OpenPositions from '../components/JoinUs/OpenPositions'
